# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationRun": {
        "additionalProperties": false,
        "properties": {
          "agent_version_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Agent Version Id",
            "type": "string"
          },
          "completed_cases": {
            "default": 0,
            "minimum": 0.0,
            "title": "Completed Cases",
            "type": "integer"
          },
          "configuration_snapshot": {
            "additionalProperties": true,
            "title": "Configuration Snapshot",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_version_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Dataset Version Id",
            "type": "string"
          },
          "evaluator_version_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Evaluator Version Ids",
            "type": "array"
          },
          "failed_cases": {
            "default": 0,
            "minimum": 0.0,
            "title": "Failed Cases",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "queued"
          },
          "total_cases": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Cases",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "agent_version_id",
          "dataset_version_id",
          "evaluator_version_ids",
          "created_at"
        ],
        "title": "EvaluationRun",
        "type": "object"
      },
      "EvaluationRunCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_version_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Agent Version Id",
            "type": "string"
          },
          "dataset_version_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Dataset Version Id",
            "type": "string"
          },
          "evaluator_version_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Evaluator Version Ids",
            "type": "array"
          }
        },
        "required": [
          "agent_version_id",
          "dataset_version_id",
          "evaluator_version_ids"
        ],
        "title": "EvaluationRunCreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegressionGateRequest": {
        "additionalProperties": false,
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RegressionGateRule"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Rules",
            "type": "array"
          }
        },
        "required": [
          "rules"
        ],
        "title": "RegressionGateRequest",
        "type": "object"
      },
      "RegressionGateResult": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RegressionGateRuleResult"
            },
            "title": "Rules",
            "type": "array"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "status": {
            "$ref": "#/components/schemas/RegressionGateStatus"
          }
        },
        "required": [
          "run_id",
          "run_status",
          "status",
          "generated_at"
        ],
        "title": "RegressionGateResult",
        "type": "object"
      },
      "RegressionGateRule": {
        "additionalProperties": false,
        "properties": {
          "aggregation": {
            "default": "pass_rate",
            "enum": [
              "average",
              "pass_rate"
            ],
            "title": "Aggregation",
            "type": "string"
          },
          "evaluator_version_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluator Version Id"
          },
          "maximum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum"
          },
          "metric_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Metric Name",
            "type": "string"
          },
          "minimum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum"
          },
          "require_all_passed": {
            "default": false,
            "title": "Require All Passed",
            "type": "boolean"
          }
        },
        "required": [
          "metric_name"
        ],
        "title": "RegressionGateRule",
        "type": "object"
      },
      "RegressionGateRuleResult": {
        "additionalProperties": false,
        "properties": {
          "actual_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Value"
          },
          "error_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Error Count",
            "type": "integer"
          },
          "failed_case_ids": {
            "items": {
              "type": "string"
            },
            "title": "Failed Case Ids",
            "type": "array"
          },
          "missing_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Missing Count",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rule": {
            "$ref": "#/components/schemas/RegressionGateRule"
          },
          "status": {
            "$ref": "#/components/schemas/RegressionGateStatus"
          },
          "valid_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Valid Count",
            "type": "integer"
          }
        },
        "required": [
          "rule",
          "status"
        ],
        "title": "RegressionGateRuleResult",
        "type": "object"
      },
      "RegressionGateStatus": {
        "enum": [
          "passed",
          "failed",
          "indeterminate",
          "incomplete"
        ],
        "title": "RegressionGateStatus",
        "type": "string"
      },
...
      "RunStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Trace-driven evaluation for prompt, RAG, tool, and custom agents.",
    "title": "Agent Eval Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{project_id}/runs": {
      "get": {
        "operationId": "list_runs_projects__project_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-project-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Key"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationRun"
                  },
                  "title": "Response List Runs Projects  Project Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "evaluation-runs"
        ]
      },
      "post": {
        "operationId": "create_run_projects__project_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-project-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Key"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "evaluation-runs"
        ]
      }
    },
...
    "/projects/{project_id}/runs/{run_id}/regression-gate": {
      "post": {
        "operationId": "evaluate_regression_gate_projects__project_id__runs__run_id__regression_gate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-project-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Key"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegressionGateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegressionGateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Regression Gate",
        "tags": [
          "regression-gates"
        ]
      }
    },
...
  }
}